# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HQuizlet Platform API
  version: 1.7.0
  description: |
    HQuizlet Platform API contract. All public application routes are exposed
    through the gateway at localhost:8080.
    Phase 9 adds File Upload: presigned URL upload flow for avatars, flashcard
    images, and study set thumbnails.
servers:
  - url: http://localhost:8080
    description: Local gateway
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ParticipantAuth:
      type: http
      scheme: bearer
      description: 'Opaque participant token issued on join. Use "Authorization: Bearer <participantToken>".'
  schemas:
    Error:
      type: object
      required: [code, message, requestId, details]
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "score must be between 0 and total"
        requestId:
          type: string
          example: "20260901143000.000000000"
        details:
          type: object
          additionalProperties: true
          example: {}
...
    StudySet:
      type: object
      required: [id, userId, title, description, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        userId: { type: integer, format: int64 }
        title: { type: string }
        description: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        thumbnailUrl: { type: [string, 'null'], description: 'Optional study set thumbnail image URL' }
        flashcards:
          type: array
          items: { $ref: '#/components/schemas/Flashcard' }
    Flashcard:
      type: object
      required: [id, studySetId, term, definition, starred, position, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        studySetId: { type: integer, format: int64 }
        term: { type: string }
        definition: { type: string }
        starred: { type: boolean }
        position: { type: integer }
        imageUrl: { type: [string, 'null'], description: 'Optional flashcard illustration image URL' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    FolderSummary:
      type: object
      required: [id, title, description, studySetCount, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        title: { type: string, minLength: 1 }
        description: { type: string }
        studySetCount: { type: integer, minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CreateFolderRequest:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
...
    StudySetList:
      type: object
      required: [items, total, page, perPage, totalPages]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/StudySet' }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        perPage: { type: integer, minimum: 1, maximum: 100 }
        totalPages: { type: integer, minimum: 0 }
...
    ClassSummary:
      type: object
      required: [id, name, description, inviteCode, memberCount, studySetCount, myRole, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, minLength: 1 }
        description: { type: string }
        inviteCode: { type: string, pattern: '^[A-Z2-9]{8}$' }
        memberCount: { type: integer, minimum: 0 }
        studySetCount: { type: integer, minimum: 0 }
        myRole: { type: string, enum: [owner, teacher, student] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ClassDetail:
      allOf:
        - { $ref: '#/components/schemas/ClassSummary' }
        - type: object
          required: [maxMembers]
          properties:
            maxMembers: { type: integer, minimum: 1, maximum: 1000 }
    CreateClassRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string }
        maxMembers: { type: integer, minimum: 2, maximum: 1000, default: 100 }
...
    ActivityItem:
      type: object
      required: [id, eventType, entityType, occurredAt]
      properties:
        id: { type: integer, format: int64 }
        eventType: { type: string }
        entityType: { type: string }
        entityId: { type: [integer, 'null'], format: int64 }
        classId: { type: [integer, 'null'], format: int64 }
        metadata: { type: object, additionalProperties: true }
        occurredAt: { type: string, format: date-time }
    ActivityFeedResponse:
      type: object
      required: [items, hasMore]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ActivityItem' }
        nextCursor: { type: string, description: 'Opaque base64 cursor for next page' }
        hasMore: { type: boolean }

    # ---- Phase 9: File Upload Schemas ----
    PresignUploadRequest:
      type: object
      required: [upload_type, filename, content_type, file_size]
      properties:
        upload_type:
          type: string
          enum: [avatar, flashcard_image, study_set_thumbnail]
        filename: { type: string, minLength: 1, maxLength: 255 }
        content_type: { type: string }
        file_size: { type: integer, format: int64, minimum: 1 }
    PresignUploadResponse:
      type: object
      required: [file_id, upload_url, upload_method, expires_at, headers_required]
      properties:
        file_id: { type: string, format: uuid }
        upload_url: { type: string, format: uri }
        upload_method: { type: string, const: PUT }
        expires_at: { type: string, format: date-time }
        headers_required:
          type: object
          properties:
            Content-Type: { type: string }
...
    FileMetadata:
      type: object
      required: [file_id, upload_type, filename, content_type, size_bytes, url, status, created_at]
      properties:
        file_id: { type: string, format: uuid }
        upload_type: { type: string }
        filename: { type: string }
        content_type: { type: string }
        size_bytes: { type: integer, format: int64 }
        url: { type: string, format: uri }
        status: { type: string, enum: [pending, active, deleted] }
        created_at: { type: string, format: date-time }
        confirmed_at: { type: [string, 'null'], format: date-time }
    QuotaInfo:
      type: object
      required: [active_files, max_files, used_bytes, max_bytes]
      properties:
        active_files: { type: integer, minimum: 0 }
        max_files: { type: integer, const: 100 }
        used_bytes: { type: integer, format: int64, minimum: 0 }
        max_bytes: { type: integer, format: int64, const: 524288000 }
    FileListResponse:
      type: object
      required: [items, total, quota]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/FileMetadata' }
        total: { type: integer, minimum: 0 }
        quota: { $ref: '#/components/schemas/QuotaInfo' }

    # ---- End Phase 9 Schemas ----

    # ---- End Phase 7 Schemas ----

    # ---- End Phase 6 Schemas ----
paths:
...
  /v1/classes:
    get:
      tags: [classes]
      summary: List classes the current user is a member of
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Class list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ClassSummary' }
        '401': { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    post:
      tags: [classes]
      summary: Create a new class
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateClassRequest' } } }
      responses:
        '201':
          description: Class created
          content: { application/json: { schema: { $ref: '#/components/schemas/ClassDetail' } } }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /v1/activity:
    get:
      tags: [activity]
      summary: Get activity feed for current user
      security: [{ BearerAuth: [] }]
      parameters:
        - { in: query, name: cursor, schema: { type: string }, description: 'Opaque cursor from previous response' }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: Activity feed
          content: { application/json: { schema: { $ref: '#/components/schemas/ActivityFeedResponse' } } }
        '401': { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ---- Phase 9: File Upload Endpoints ----
  /v1/files/presign:
    post:
      tags: [files]
      summary: Request a presigned URL for file upload
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PresignUploadRequest' }
      responses:
        '201':
          description: Presigned URL created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PresignUploadResponse' }
        '400':
          description: Invalid request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' } 
        '401': { description: Authentication required }
        '429':
          description: Quota exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /v1/files:
    get:
      tags: [files]
      summary: List user files with quota info
      security: [{ BearerAuth: [] }]
      parameters:
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: per_page, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: File list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileListResponse' }
        '401': { description: Authentication required }

  # ---- End Phase 9 Endpoints ----

  # ---- End Phase 7 Endpoints ----
...
  /v1/study-sets:
    get:
      tags: [study-sets]
      security: [{ BearerAuth: [] }]
      parameters:
        - { in: query, name: search, schema: { type: string } }
        - { in: query, name: sort, schema: { type: string, enum: [updated, created, title] } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: per_page, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { description: Paginated study sets, content: { application/json: { schema: { $ref: '#/components/schemas/StudySetList' } } } }
    post:
      tags: [study-sets]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/StudySet' } } } }
...
  /v1/folders:
    get:
      tags: [folders]
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Folders, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/FolderSummary' } } } } }
    post:
      tags: [folders]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderRequest'
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/FolderSummary' } } } }
        '422': { description: Invalid payload, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...